splits servlet into two for reading and writing; builds game object when serving objects instead of storing
